moved file and added CSS
